+++ src/epomakercontroller/utils/sensors.py
@@ -7,7 +7,6 @@ def get_cpu_usage(test_mode: bool = False) -> int:
     if test_mode:
         return random.randint(0, 99)
     return int(round(psutil.cpu_percent(interval=1)))
-    # click.echo(f"CPU Usage: {cpu_usage}%, sending {cpu_usage_rounded}%")
 
 
 def get_device_temp(temp_key: str | None, test_mode: bool = False) -> int | None:
